Introduce simple mode to login procedure
When token needs to be updated, the location actually is not
necessarily to be updated.
In simple mode, location selection prompts  output
only when a phone number section does not exist in credential ini file.
In another words, if a phone number exists and the login process is
called with -s/--simple-mode option,
the only thing after input the phone number is to input the
verification code.